Create a SignUpPage component
